requirements.txt
aenum==3.1.12
anyio==3.6.2
argcomplete==3.0.5
async-timeout==4.0.2
asyncio-mqtt==0.16.1
asyncpg==0.27.0
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
datamodel-code-generator==0.18.0
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
filip==0.2.5
fonttools==4.39.3
genson==1.2.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
igraph==0.9.8
importlib-resources==5.12.0
inflect==5.6.2
isodate==0.6.1
isort==5.12.0
Jinja2==3.1.2
jsonpath-ng==1.5.3
jsonschema==4.17.3
jsonschema-spec==0.1.4
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.2
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.0
pandas-datapackage-reader==0.18.0
pathable==0.4.3
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
ply==3.11
prance==0.22.2.22.0
psycopg2-binary==2.9.6
pydantic==1.10.7
pyparsing==3.0.9
pyrsistent==0.19.3
PySnooper==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
rdflib==6.0.2
redis==4.5.4
regex==2023.3.23
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
stringcase==1.2.0
texttable==1.6.7
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
wget==3.2
